Move helper functions to pretranslate.py Rename utils.py to tasks.py Change all variants to pretrans* (which thought me to be consistent with names by giving a real hard time to fix)
